Add PDDL problems and TestCase interface
